VAKT G4P Sjekk om periodeIder er null (#2821)
Legg på null sjekk for å unngå `NullPointerException`
https://nav-it.slack.com/archives/C01BSCJM127/p1747131806703339

+++ service/src/main/java/no/nav/melosys/service/behandling/BehandlingService.java
@@ -155,6 +155,12 @@ public class BehandlingService {
         if (!behandling.erAktiv()) {
             throw new FunksjonellException("Medlemsperioder kan ikke lagres på behandling med status " + behandling.getStatus());
         }
+        if(periodeIder == null) {
+            // Det er en funksjonell feil og brurde valideres og gi tilbake melding til bruker
+            // Men for å hindre støy i #melosys-alarmer-prod frem til det blir fikset så logger vi det som en warning
+            log.warn("Kan ikke knytte medlemsperioder til behandling {} fordi periodeIder er null", behandlingID);
+            return;
+        }
         List<TidligereMedlemsperiode> tidligereMedlemsperioder = periodeIder.stream()
             .map(pid -> new TidligereMedlemsperiode(behandlingID, pid)).toList();
         tidligereMedlemsperiodeRepository.deleteById_BehandlingId(behandlingID);
